Standardize UnfrozenConfig and FlatConfig classes
- Move aliases to Config class and enables `min_verbosity` for
UnfrozenConfigs(missed in previous refactor)
- Hide alias_mods from dump for UnfrozenConfig
